requirements.txt
asgiref==3.4.1
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.10.0
django-heroku==0.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
flake8==3.9.2
gunicorn==20.1.0
mccabe==0.6.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==5.3.0